ci: use .scripts/ls.py in .github/get-packages.py
